fix: config updates
