+++ package.json
@@ -63,10 +63,17 @@
   "scripts": {
     "setup": "bash scripts/setup.sh",
     "clean": "rm -rf dist node_modules/.bun site/node_modules",
-    "build": "bun run build:cjs && bun run build:esm && bun run build:core && bun run build:cli && bun run build:types",
+    "build": "bun run build:clean && bun run build:cjs && bun run build:esm && bun run build:core && bun run build:exports && bun run build:cli && bun run build:types",
+    "build:clean": "rm -rf dist",
     "build:cjs": "bun build src/index.ts --outfile dist/index.cjs --format cjs --minify",
     "build:esm": "bun build src/index.ts --outfile dist/index.mjs --format esm --minify",
     "build:core": "bun build src/core.ts --outfile dist/core.mjs --format esm --minify && bun build src/core.ts --outfile dist/core.cjs --format cjs --minify",
+    "build:exports": "bun run build:themes && bun run build:theme-presets && bun run build:tokenizer && bun run build:renderer && bun run build:fast",
+    "build:themes": "bun build src/themes/index.ts --outfile dist/themes/index.mjs --format esm --minify && bun build src/themes/index.ts --outfile dist/themes/index.cjs --format cjs --minify",
+    "build:theme-presets": "bun build src/themes/presets/*.ts --outdir dist/themes/presets --entry-naming '[name].mjs' --format esm --minify && bun build src/themes/presets/*.ts --outdir dist/themes/presets --entry-naming '[name].cjs' --format cjs --minify",
+    "build:tokenizer": "bun build src/tokenizer/index.ts --outfile dist/tokenizer/index.mjs --format esm --minify && bun build src/tokenizer/index.ts --outfile dist/tokenizer/index.cjs --format cjs --minify",
+    "build:renderer": "bun build src/renderer/index.ts --outfile dist/renderer/index.mjs --format esm --minify && bun build src/renderer/index.ts --outfile dist/renderer/index.cjs --format cjs --minify",
+    "build:fast": "bun build src/renderer/fast-mode.ts --outfile dist/renderer/fast-mode.mjs --format esm --minify && bun build src/renderer/fast-mode.ts --outfile dist/renderer/fast-mode.cjs --format cjs --minify",
     "build:cli": "bun build src/cli/bin.ts --outfile dist/cli.js --format cjs --minify --target node",
     "build:types": "tsc -p tsconfig.build.json",
     "test": "bun test tests/",
@@ -87,6 +94,7 @@
     "site:start": "cd site && bun run start",
     "dev": "mprocs",
     "create-theme": "bun run src/cli/index.ts theme create",
+    "prepack": "bun run build",
     "postinstall": "bun run scripts/install-hooks.ts"
   },
   "homepage": "https://jeffry.in/logsdx",
