+++ src/cli/commands.ts
@@ -1,6 +1,6 @@
 import { input, select, checkbox, confirm } from "../utils/prompts";
 import spinner from "../utils/spinner";
-import * as colorUtil from "../utils/colors";
+import chalk from "../utils/colors";
 import gradient from "../utils/gradient";
 import boxen from "../utils/boxen";
 import { writeFileSync, existsSync, mkdirSync } from "fs";
@@ -14,87 +14,36 @@ import {
 import { registerTheme, getTheme } from "../themes";
 import { getLogsDX } from "../index";
 import { Theme } from "../types";
+import {
+  SAMPLE_LOGS,
+  COLOR_PRESETS,
+  BANNER_ASCII,
+  THEME_CREATOR_VERSION,
+  PROMPTS,
+  THEME_MODES,
+  FEATURE_PRESETS,
+  SAVE_OPTIONS,
+} from "./constants";
+import { createLogger } from "../utils/logger";
 
-const SAMPLE_LOGS = [
-  "INFO: Server started on port 3000",
-  "WARN: Memory usage high: 85%",
-  "ERROR: Database connection failed",
-  "DEBUG: Processing request id=12345",
-  "SUCCESS: All tests passed ✓",
-  "[2024-01-15 10:23:45] Request completed",
-  "GET /api/users 200 OK (123ms)",
-  "Cache hit ratio: 92.5%",
-];
-
-const COLOR_PRESETS = {
-  Vibrant: {
-    primary: "#007acc",
-    error: "#ff4444",
-    warning: "#ff9900",
-    success: "#00cc66",
-    info: "#00aaff",
-    muted: "#666666",
-  },
-  Pastel: {
-    primary: "#87ceeb",
-    error: "#ffb6c1",
-    warning: "#ffd700",
-    success: "#98fb98",
-    info: "#add8e6",
-    muted: "#d3d3d3",
-  },
-  Neon: {
-    primary: "#00ffff",
-    error: "#ff0080",
-    warning: "#ffff00",
-    success: "#00ff00",
-    info: "#ff00ff",
-    muted: "#808080",
-  },
-  Earth: {
-    primary: "#8b7355",
-    error: "#cd5c5c",
-    warning: "#daa520",
-    success: "#228b22",
-    info: "#4682b4",
-    muted: "#696969",
-  },
-  Ocean: {
-    primary: "#006994",
-    error: "#ff6b6b",
-    warning: "#ffd93d",
-    success: "#4ecdc4",
-    info: "#45b7d1",
-    muted: "#95a5a6",
-  },
-  Custom: null,
-};
+const log = createLogger("commands");
 
 function showBanner() {
-  console.clear();
   const grad = gradient();
-  console.log(
-    grad.multiline(`
-  ╦  ┌─┐┌─┐┌─┐╔╦╗═╗ ╦
-  ║  │ ││ ┬└─┐ ║║╔╩╦╝
-  ╩═╝└─┘└─┘└─┘═╩╝╩ ╚═
-  `),
-  );
-  console.log(colorUtil.dim("  Theme Creator v1.0.0\n"));
+  log.print(grad.multiline(BANNER_ASCII));
+  log.print(chalk.dim(`  ${THEME_CREATOR_VERSION}\n`));
 }
 
 async function renderPreview(theme: Theme, title: string = "Theme Preview") {
   const logsDX = await getLogsDX({ theme, outputFormat: "ansi" });
-  const previewBox = boxen(
-    SAMPLE_LOGS.map((log) => logsDX.processLine(log)).join("\n"),
-    {
-      title,
-      padding: 1,
-      borderStyle: "round",
-      borderColor: "cyan",
-    },
-  );
-  console.log(previewBox);
+  const styledLogs = SAMPLE_LOGS.map((line) => logsDX.processLine(line));
+  const previewBox = boxen(styledLogs.join("\n"), {
+    title,
+    padding: 1,
+    borderStyle: "round",
+    borderColor: "cyan",
+  });
+  log.print(previewBox);
 }
 
 export async function createInteractiveTheme(
@@ -105,12 +54,12 @@ export async function createInteractiveTheme(
   }
 
   const name = await input({
-    message: "Theme name:",
+    message: PROMPTS.themeName,
     validate: async (inputValue: string) => {
-      if (!inputValue.trim()) return "Theme name is required";
+      if (!inputValue.trim()) return PROMPTS.themeNameRequired;
       try {
         await getTheme(inputValue);
-        return "A theme with this name already exists";
+        return PROMPTS.themeExists;
       } catch {
         return true;
       }
@@ -120,17 +69,13 @@ export async function createInteractiveTheme(
   });
 
   const description = await input({
-    message: "Theme description:",
+    message: PROMPTS.themeDescription,
     default: "",
   });
 
   const mode = await select({
-    message: "Theme mode:",
-    choices: [
-      { name: "Dark (for dark terminals)", value: "dark" },
-      { name: "Light (for light terminals)", value: "light" },
-      { name: "Auto (system preference)", value: "auto" },
-    ],
+    message: PROMPTS.themeMode,
+    choices: THEME_MODES.map((m) => ({ name: m.name, value: m.value })),
     default: "dark",
   });
 
@@ -141,116 +86,102 @@ export async function createInteractiveTheme(
   };
 
   const preset = await select({
-    message: "Choose a color preset:",
+    message: PROMPTS.chooseColorPreset,
     choices: Object.keys(COLOR_PRESETS).map((name) => ({
-      name: name === "Custom" ? "Custom (define your own)" : name,
+      name: name === "Custom" ? PROMPTS.customDefineOwn : name,
       value: name,
     })),
   });
 
-  let colors = COLOR_PRESETS[preset as keyof typeof COLOR_PRESETS];
+  let themeColors: SimpleThemeConfig["colors"] | null =
+    COLOR_PRESETS[preset as keyof typeof COLOR_PRESETS];
 
-  if (preset === "Custom" || !colors) {
-    const loadingSpinner = spinner("Loading color picker...").start();
+  if (preset === "Custom" || !themeColors) {
+    const loadingSpinner = spinner(PROMPTS.loadingColorPicker).start();
     await new Promise((r) => setTimeout(r, 500));
     loadingSpinner.stop();
 
+    const validateHex = (inputValue: string) =>
+      /^#[0-9A-Fa-f]{6}$/.test(inputValue) || PROMPTS.invalidHexColor;
+
     const primary = await input({
-      message: "Primary color (hex):",
+      message: PROMPTS.primaryColor,
       default: "#ffffff",
-      validate: (inputValue: string) =>
-        /^#[0-9A-Fa-f]{6}$/.test(inputValue) || "Invalid hex color",
-      transformer: (inputValue: string) =>
-        colorUtil.cyan(inputValue || "#ffffff"),
+      validate: validateHex,
+      transformer: (inputValue: string) => chalk.cyan(inputValue || "#ffffff"),
     });
 
     const error = await input({
-      message: "Error color (hex):",
+      message: PROMPTS.errorColor,
       default: "#ff4444",
-      validate: (inputValue: string) =>
-        /^#[0-9A-Fa-f]{6}$/.test(inputValue) || "Invalid hex color",
-      transformer: (inputValue: string) =>
-        colorUtil.red(inputValue || "#ff4444"),
+      validate: validateHex,
+      transformer: (inputValue: string) => chalk.red(inputValue || "#ff4444"),
     });
 
     const warning = await input({
-      message: "Warning color (hex):",
+      message: PROMPTS.warningColor,
       default: "#ff9900",
-      validate: (inputValue: string) =>
-        /^#[0-9A-Fa-f]{6}$/.test(inputValue) || "Invalid hex color",
+      validate: validateHex,
       transformer: (inputValue: string) =>
-        colorUtil.yellow(inputValue || "#ff9900"),
+        chalk.yellow(inputValue || "#ff9900"),
     });
 
     const success = await input({
-      message: "Success color (hex):",
+      message: PROMPTS.successColor,
       default: "#00cc66",
-      validate: (inputValue: string) =>
-        /^#[0-9A-Fa-f]{6}$/.test(inputValue) || "Invalid hex color",
-      transformer: (inputValue: string) =>
-        colorUtil.green(inputValue || "#00cc66"),
+      validate: validateHex,
+      transformer: (inputValue: string) => chalk.green(inputValue || "#00cc66"),
     });
 
     const info = await input({
-      message: "Info color (hex):",
+      message: PROMPTS.infoColor,
       default: "#00aaff",
-      validate: (inputValue: string) =>
-        /^#[0-9A-Fa-f]{6}$/.test(inputValue) || "Invalid hex color",
-      transformer: (inputValue: string) =>
-        colorUtil.blue(inputValue || "#00aaff"),
+      validate: validateHex,
+      transformer: (inputValue: string) => chalk.blue(inputValue || "#00aaff"),
     });
 
     const muted = await input({
-      message: "Muted color (hex):",
+      message: PROMPTS.mutedColor,
       default: "#666666",
-      validate: (inputValue: string) =>
-        /^#[0-9A-Fa-f]{6}$/.test(inputValue) || "Invalid hex color",
-      transformer: (inputValue: string) =>
-        colorUtil.gray(inputValue || "#666666"),
+      validate: validateHex,
+      transformer: (inputValue: string) => chalk.gray(inputValue || "#666666"),
     });
 
-    colors = { primary, error, warning, success, info, muted };
+    themeColors = { primary, error, warning, success, info, muted };
   }
 
   const presets = await checkbox({
-    message: "Select features to highlight:",
-    choices: [
-      {
-        name: "Log levels (ERROR, WARN, INFO)",
-        value: "logLevels",
-        checked: true,
-      },
-      { name: "Numbers and numeric values", value: "numbers", checked: true },
-      { name: "Dates and timestamps", value: "dates", checked: true },
-      { name: "Boolean values", value: "booleans", checked: true },
-      { name: "Brackets and punctuation", value: "brackets", checked: true },
-      { name: "Quoted strings", value: "strings", checked: false },
-    ],
+    message: PROMPTS.selectFeatures,
+    choices: FEATURE_PRESETS.map((f) => ({
+      name: f.name,
+      value: f.value,
+      checked: f.checked,
+    })),
   });
 
-  const createSpinner = spinner("Creating theme...").start();
+  const createSpinner = spinner(PROMPTS.creatingTheme).start();
 
   const config: SimpleThemeConfig = {
     name: basicInfo.name,
     description: basicInfo.description || undefined,
     mode: basicInfo.mode as "light" | "dark" | "auto",
-    colors: colors!,
+    colors: themeColors!,
     presets,
   };
 
   const theme = createTheme(config);
-  createSpinner.succeed("Theme created!");
+  createSpinner.succeed(PROMPTS.themeCreated);
 
-  console.log("\n");
+  log.print("\n");
   await renderPreview(theme, `${theme.name} Preview`);
 
   const checkAccessibility = await confirm({
-    message: "Check accessibility compliance?",
+    message: PROMPTS.checkAccessibility,
     default: true,
   });
 
   if (checkAccessibility) {
-    const accessSpinner = spinner("Checking accessibility...").start();
+    const accessSpinner = spinner(PROMPTS.checkingAccessibility).start();
     const result = checkWCAGCompliance(theme);
     accessSpinner.stop();
 
@@ -260,9 +191,9 @@ export async function createInteractiveTheme(
         (result.recommendations.length > 0
           ? "\nRecommendations:\n" +
             result.recommendations.map((r: string) => `- ${r}`).join("\n")
-          : "\nNo issues found"),
+          : `\n${PROMPTS.noIssuesFound}`),
       {
-        title: "Accessibility Report",
+        title: PROMPTS.accessibilityReport,
         padding: 1,
         borderStyle: "round",
         borderColor:
@@ -273,51 +204,42 @@ export async function createInteractiveTheme(
               : "green",
       },
     );
-    console.log(accessBox);
+    log.print(accessBox);
 
     if (result.recommendations.length > 0) {
       const fixIssues = await confirm({
-        message: "Auto-fix accessibility issues?",
+        message: PROMPTS.autoFixAccessibility,
         default: true,
       });
 
       if (fixIssues) {
-        const fixSpinner = spinner("Fixing accessibility issues...").start();
+        const fixSpinner = spinner(PROMPTS.fixingAccessibility).start();
         const fixedTheme = adjustThemeForAccessibility(theme, 4.5);
-        fixSpinner.succeed("Accessibility issues fixed!");
+        fixSpinner.succeed(PROMPTS.accessibilityFixed);
         Object.assign(theme, fixedTheme);
       }
     }
   }
 
   const saveOption = await select({
-    message: "How would you like to save the theme?",
-    choices: [
-      { name: "Export as JSON file", value: "json" },
-      { name: "Export as TypeScript file", value: "typescript" },
-      { name: "Copy to clipboard", value: "clipboard" },
-      { name: "Register for immediate use", value: "register" },
-      { name: "Don't save", value: "none" },
-    ],
+    message: PROMPTS.saveThemeHow,
+    choices: SAVE_OPTIONS.map((o) => ({ name: o.name, value: o.value })),
   });
 
   if (saveOption !== "none") {
     await saveTheme(theme, saveOption);
   }
 
-  console.log(
-    boxen(
-      colorUtil.green("Theme creation complete!\n\n") +
-        colorUtil.dim(
-          `Use your theme with: ${colorUtil.cyan(`logsdx --theme ${theme.name}`)}`,
-        ),
-      {
-        padding: 1,
-        borderStyle: "round",
-        borderColor: "green",
-      },
-    ),
+  const completionMsg = chalk.green(`${PROMPTS.themeCreationComplete}\n\n`);
+  const usageHint = chalk.dim(
+    `Use your theme with: ${chalk.cyan(`logsdx --theme ${theme.name}`)}`,
   );
+  const completionBox = boxen(completionMsg + usageHint, {
+    padding: 1,
+    borderStyle: "round",
+    borderColor: "green",
+  });
+  log.print(completionBox);
 }
 
 async function saveTheme(theme: Theme, saveOption: string) {
@@ -335,7 +257,7 @@ async function saveTheme(theme: Theme, saveOption: string) {
 
   if (saveOption === "json") {
     const filepath = await input({
-      message: "Save as:",
+      message: PROMPTS.saveAs,
       default: `./themes/${theme.name}.json`,
     });
 
@@ -345,10 +267,10 @@ async function saveTheme(theme: Theme, saveOption: string) {
     }
 
     writeFileSync(filepath, JSON.stringify(themeData, null, 2));
-    console.log(colorUtil.green(`Saved to ${filepath}`));
+    log.success(`Saved to ${filepath}`);
   } else if (saveOption === "typescript") {
     const filepath = await input({
-      message: "Save as:",
+      message: PROMPTS.saveAs,
       default: `./themes/${theme.name}.ts`,
     });
 
@@ -363,6 +285,6 @@ export const ${theme.name.replace(/[^a-zA-Z0-9]/g, "_")}Theme: Theme = ${JSON.st
 `;
 
     writeFileSync(filepath, tsContent);
-    console.log(colorUtil.green(`Saved to ${filepath}`));
+    log.success(`Saved to ${filepath}`);
   }
 }
