+++ src/schema/utils.ts
@@ -1,9 +1,4 @@
 import { COLOR_PATTERN } from "./constants";
-import {
-  ValidationError,
-  isValidationError,
-  formatValidationIssues,
-} from "../lib/validate";
 
 export function isValidColorFormat(color: string): boolean {
   return COLOR_PATTERN.test(color);
@@ -14,5 +9,3 @@ export function createValidationError(message: string, cause: Error): Error {
   error.cause = cause;
   return error;
 }
-
-export { isValidationError, formatValidationIssues, ValidationError };
