--- src/tokenizer/cache-constants.ts
@@ -1,2 +1,0 @@
-export const MAX_CACHE_SIZE = 10;
-export const CACHE_TTL = 60000; // 1 minute
