+++ src/tokenizer/types.ts
@@ -1,3 +1,5 @@
+import type { SimpleLexer } from "./index";
+
 export type MatcherType =
   | "word"
   | "regex"
@@ -12,3 +14,13 @@ export type MatcherType =
   | "newline"
   | "carriage-return"
   | "default";
+
+export interface CachedLexer {
+  lexer: SimpleLexer;
+  themeHash: string;
+  lastUsed: number;
+}
+
+export interface CacheOptions {
+  trim?: string;
+}
