+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 export { default as spinner } from "./spinner";
-export type { Spinner } from "./spinner";
+export type { Spinner } from "./types";
 export { default as colors } from "./colors";
 export { default as gradient } from "./gradient";
 export { default as ascii } from "./ascii";
