--- src/utils/strip-ansi.ts
@@ -1,8 +1,0 @@
-const ansiRegex =
-  /[\u001B\u009B][[\]()#;?]*(?:(?:(?:(?:;[-a-zA-Z\d/#&.:=?%@~_]+)*|[a-zA-Z\d]+(?:;[-a-zA-Z\d/#&.:=?%@~_]*)*)?\u0007)|(?:(?:\d{1,4}(?:;\d{0,4})*)?[\dA-PR-TZcf-ntqry=><~]))/g;
-
-export function stripAnsi(str: string): string {
-  return str.replace(ansiRegex, "");
-}
-
-export default stripAnsi;
